refactor:useDiscount훅에 함수 추가
